% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignFunctionInPackage.R
\name{assignFunctionInPackage}
\alias{assignFunctionInPackage}
\title{Overwrite functions inside R packages}
\usage{
assignFunctionInPackage(fun, name, package)
}
\arguments{
\item{fun}{: function you want to replace}

\item{name}{: the name of the function in a package}

\item{package:}{the name of the package

example usage:
avoid a parallel backend from being registered which triggers a firewall message}
}
\description{
Overwrite functions inside R packages
}
