% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformBoxCox.R
\name{BoxCox}
\alias{BoxCox}
\alias{invBoxCox}
\alias{BoxCoxAuto}
\alias{BoxCoxLambda}
\title{Box-Cox transformation}
\usage{
BoxCox(x, lambda, shift = 0)

invBoxCox(x, lambda, shift = 0)

BoxCoxAuto(
  x,
  method = c("loglik", "minitab"),
  lower = -5,
  upper = 5,
  shift = 0
)

BoxCoxLambda(
  x,
  method = c("loglik", "minitab"),
  lower = -5,
  upper = 5,
  shift = 0
)
}
\arguments{
\item{x}{a numeric vector.}

\item{lambda}{The lambda parameter of the transform.}

\item{shift}{The shift parameter of the transform. \code{shift=0} by default. All values of \code{x} must be larger than \code{shift}.}

\item{method}{Choose method to be used in calculating lambda. See details for explanation.}

\item{lower}{Lower limit for possible lambda values.}

\item{upper}{Upper limit for possible lambda values.}
}
\value{
\code{BoxCox} and \code{BoxCoxAuto} return the transformed variable. \code{invBoxCox} returns the untransformed variable.
\code{BoxCoxLambda} returns the optimal value of \code{lambda} for the given data.
}
\description{
Transform a variable using the two-parameter Box-Cox transformation.
It is possible to select the value of \code{lambda} automatically
using one of two available methods.
}
\details{
If \code{method=="loglik"}, the value of lambda is chosen to maximize the
normal profile log likelihood of the transformed variable.

If \code{method=="minitab"}, the value of lambda is chosen to minimize
the sample standard deviation of the standardized, Box-Cox transformed variable.
This behavior mimics the auto Box-Cox transform available in the Minitab software.
}
\references{
Box, G. E. P. and Cox, D. R. (1964) An analysis of
transformations. \emph{JRSS B} \bold{26} 211--246.

https://support.minitab.com/en-us/minitab/help-and-how-to/quality-and-process-improvement/control-charts/how-to/box-cox-transformation/methods-and-formulas/methods-and-formulas/
}
