% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kknn-contr.R
\name{contrasts}
\alias{contrasts}
\alias{contr.dummy}
\alias{contr.ordinal}
\alias{contr.metric}
\title{Contrast matrices}
\value{
A matrix of contrasts.
}
\description{
These functions are re-exports of \code{\link[kknn:contr.dummy]{kknn::contr.dummy()}}, \code{\link[kknn:contr.dummy]{kknn::contr.ordinal()}}, and \code{\link[kknn:contr.dummy]{kknn::contr.metric()}}.
}
\keyword{internal}
