% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{excludeNaListwise}
\alias{excludeNaListwise}
\title{Exclude rows with missing values (listwise deletion)}
\usage{
excludeNaListwise(dataset, columns = NULL)
}
\arguments{
\item{dataset}{dataframe containing the dataset.}

\item{columns}{a character vector with column names, or NULL to remove all rows with missing values.}
}
\value{
a dataframe with rows that contain missing values removed
}
\description{
Exclude rows with missing values (listwise deletion)
}
