% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleInstall.R
\name{getModuleData}
\alias{getModuleData}
\title{Extract Relevant Data from a JASP Module}
\usage{
getModuleData(jaspRoot, modulePath, recurseJaspDependencies)
}
\arguments{
\item{jaspRoot}{path to root of local jasp-desktop clone}

\item{modulePath}{path to module to install}

\item{recurseJaspDependencies}{whether to recursively find jasp module dependencies}
}
\value{
A list with the following named objects:
\itemize{
\item \code{cachedObject} A cached object with previously computed results. Only exists when building JASP.
\item \code{localPaths} The local paths to the to be installed modules.
\item \code{commitHashes} All R package dependencies of the module.
\item \code{jaspPkgs} All JASP Module dependencies.
}
}
\description{
Extract Relevant Data from a JASP Module
}
