% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOrRetrieve.R
\name{isRecomputed}
\alias{isRecomputed}
\title{Test if a jaspObject was recomputed of not.}
\usage{
isRecomputed(x)
}
\arguments{
\item{x}{a jaspObject, or missing in which case the last created or retrieved jaspObject is considered.}
}
\description{
Test if a jaspObject was recomputed of not.
}
\details{
Note that this function only works when \code{\link{\%setOrRetrieve\%}} is used to set or retrieve a jaspObject.
}
