% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspBase.R
\docType{package}
\name{jaspBase-package}
\alias{jaspBase}
\alias{jaspBase-package}
\title{jaspBase: A package for interfacing between Jasp and R}
\description{
Package contains the JASP Bayesian and Frequentist analyses. See \url{https://jasp-stats.org} for more details.
}
