% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspDeps.R
\name{jaspDeps}
\alias{jaspDeps}
\title{jaspDeps - An object to store jaspDependencies}
\usage{
jaspDeps(
  options = NULL,
  optionsFromObject = NULL,
  optionContainsValue = NULL,
  nestedOptions = NULL,
  nestedOptionsContainsValue = NULL
)
}
\arguments{
\item{options}{passed to \verb{$dependOn}.}

\item{optionsFromObject}{passed to \verb{$dependOn}.}

\item{optionContainsValue}{passed to \verb{$dependOn}.}

\item{nestedOptions}{passed to \verb{$dependOn}.}

\item{nestedOptionsContainsValue}{passed to \verb{$dependOn}.}
}
\description{
jaspDeps - An object to store jaspDependencies
}
\details{
The function is a convenience wrapper to set dependencies in the constructor of jaspObjects.
For example, this code
\preformatted{
container <- createJaspContainer(title = "Title")
container$dependOn(optionContainsValue = list(variables = "contNormal"))
}
is identical to
\preformatted{
container <- createJaspContainer(
  title = "Title",
  dependencies = jaspDeps(optionContainsValue = list(variables = "contNormal"))
)
}
}
