% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformJohnson.R
\name{Johnson}
\alias{Johnson}
\title{Johnson transformation}
\usage{
Johnson(x, lower = 0.25, upper = 1.25)
}
\arguments{
\item{x}{a numeric vector to be transformed.}

\item{lower}{Lower limit for possible z values.}

\item{upper}{Upper limit for possible z values.}
}
\value{
Numeric vector of the transformed values. For convenience, the transformation type and parameters are stored in the object's attributes.
}
\description{
Transform a variable using the Johnson transformation that transforms unbounded variables into a close-to-normal
distribution. It is computationally more expensive than \link{BoxCox}, \link{powerTransform}, or \link{YeoJohnson}, but also more
expressive, allowing better normalization for more complicated empirical distributions.
}
\details{
This transformation follows the algorithm described in Chou, et al. (1998), with the exception that instead of
calculating the Shapiro-Wilk test statistic W for every \eqn{z\in\{0.25, 0.26, \dots, 1.24, 1.25\}},
we instead use \link[stats:optimize]{stats::optimise} to find z that gives the largest W.
}
\references{
Chou, Y. M., Polansky, A. M., & Mason, R. L. (1998). Transforming non-normal data to normality in statistical process control. \emph{Journal of Quality Technology, 30}(2), 133-141.
}
