% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathExpression.R
\name{mathExpression}
\alias{mathExpression}
\title{LaTeX Expressions in JASP}
\usage{
mathExpression(x, inline = TRUE)
}
\arguments{
\item{x}{A string that contains the LaTeX code for the math expression}

\item{inline}{Should the expression be inline or in display mode?}
}
\description{
Construct LaTeX expressions for JASP output
}
\examples{
# Escape the backslash character for LaTeX code
jaspBase::mathExpression("\\\\alpha + \\\\beta")
# Alternatively use the raw character constant
jaspBase::mathExpression(r"{\alpha + \beta}")

# By default expressions are inline, to disable use
jaspBase::mathExpression(r"{ \bar{x} = \frac{\sum_i^n x_i}{n} }", inline = FALSE)

}
