% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformPower.R
\name{powerTransform}
\alias{powerTransform}
\alias{powerTransformAuto}
\alias{powerTransformLambda}
\title{Power transformation transformation}
\usage{
powerTransform(x, lambda, shift = 0)

powerTransformAuto(x, predictor = NULL, lower = -5, upper = 5, shift = 0)

powerTransformLambda(x, predictor = NULL, lower = -5, upper = 5, shift = 0)
}
\arguments{
\item{x}{a numeric vector.}

\item{lambda}{The lambda parameter of the transform.}

\item{shift}{The shift parameter of the transform. \code{shift=0} by default. Values of \code{x} smaller than \code{shift} will turn \code{NA}.}

\item{predictor}{Vector indicating regression variable (factor or numeric).}

\item{lower}{Lower limit for possible lambda values.}

\item{upper}{Upper limit for possible lambda values.}
}
\value{
\code{powerTransform} and \code{powerTransformAuto} return the transformed variable.
\code{powerTransformLambda} returns the optimal value of \code{lambda} for the given data.
}
\description{
Transform a variable using the power transformation (scale-invariant Box-Cox).
It is possible to select the value of \code{lambda} automatically.
}
\details{
The power transform is defined as
\deqn{y_i =
\begin{cases}
\frac{(x_i+\text{shift})^\lambda - 1}{\lambda \mathrm{GM}(\mathbb{x+\text{shift}})^{\lambda-1}} & \mathrm{if } \lambda \neq 0 \\
\mathrm{GM}(\mathbb{x+\text{shift}})\log(x_i+\text{shift}) & \mathrm{if } \lambda = 0,
\end{cases}
}
where \eqn{\mathrm{GM}(\mathbb{x})} is the geometric mean of observations \eqn{x = x_1, \dots, x_n}.

Automatic transform finds lambda that minimizes the residual sums of squares of the power-transformed variable
regressed on the predictor (Box & Cox, 1964, p. 216).
}
\references{
Box, G. E. P. and Cox, D. R. (1964) An analysis of transformations. \emph{JRSS B} \bold{26} 211--246.
}
