% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{readDataSetByVariableTypes}
\alias{readDataSetByVariableTypes}
\title{readDataSetByVariableTypes}
\usage{
readDataSetByVariableTypes(options, keys, exclude.na.listwise = NULL)
}
\arguments{
\item{options}{options from QML.}

\item{keys}{character, option name(s) that contain variables in the dataset.}

\item{exclude.na.listwise}{character, column names for which any missing values will cause that row to be excluded.}
}
\description{
readDataSetByVariableTypes
}
\details{
\code{readDataSetByVariableTypes} automatically removes keys that are empty lists or empty strings, unlike \code{.readDataSetToEnd} which would throw an error.
}
