% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOrRetrieve.R
\name{\%setOrRetrieve\%}
\alias{\%setOrRetrieve\%}
\title{Set or retrieve a jaspObject}
\usage{
lhs \%setOrRetrieve\% rhs
}
\arguments{
\item{lhs}{an assignment into a jaspObject, e.g., \code{container[["table"]]}.}

\item{rhs}{a function call that creates a jaspObject.}
}
\description{
Set or retrieve a jaspObject
}
\details{
This function exists as a shorthand for the following very common pattern:
\preformatted{
if (is.null(jaspContainer[[key]])) { # was this subelement already computed?
  subContainer <- createJaspContainer(..., dependencies = ...) # no, so recreate it
  jaspContainer[[key]] <- subContainer # store it in this position
} else {
  subContainer <- jaspContainer[[key]] # it was recomputed, retrieve it from the state.
}
}
The code above duplicates the phrase \code{jaspContainer[[key]]} quite a bit.
When this is a string literal, this introduces a lot of room for copy-paste errors.
with \verb{\%setOrRetrieve\%}, this becomes
\preformatted{
subContainer <- jaspContainer[[key]] \%setOrRetrieve\% createJaspContainer(..., dependencies = ...)
}

The same pattern can also be used to set and retrieve state objects.
Consider the following code
\preformatted{
if (is.null(jaspContainer[[key]])) {
  object <- expensiveComputeFunction()
  state <- createJaspState(object, dependencies = ...)
  jaspContainer[[key]] <- state
} else {
  object <- jaspContainer[[key]]$object
}
}
with \verb{\%setOrRetrieve\%}, this becomes
\preformatted{
object <- jaspContainer[[key]] \%setOrRetrieve\% (
  expensiveComputeFunction() |>
  createJaspState(dependencies = ...)
)
}
Note that if the \code{rhs} passed to \verb{\%setOrRetrieve\%} returns an object of class \code{jaspStateR} then \verb{\%setOrRetrieve\%} returns the object it contains, rather than the jaspObject itself.
In all other cases the jaspObject is returned.
If the \code{rhs} does not return a jaspObject, an error is thrown.
}
\examples{
library(jaspBase)

# otherwise jaspState crashes badly (should be fixed another time)
jaspResultsCPP        <- jaspBase:::loadJaspResults("a name")
jaspResultsCPP$title  <- "a title"
jaspResults           <- jaspBase:::jaspResultsR$new(jaspResultsCPP)

ctr <- createJaspContainer("ctr1")

result <- ctr[["ctr2"]] \%setOrRetrieve\%
  createJaspContainer("my name")

result               # the container
isRecomputed(result) # true
isRecomputed()       # true

# the same statement as before
result <- ctr[["ctr2"]] \%setOrRetrieve\%
  createJaspContainer("my name")

isRecomputed(result) # false
isRecomputed()       # false

set.seed(123)
computeFunction <- function() {
  print("working hard!")
  c(a = rnorm(1))
}

result <- ctr[["state"]] \%setOrRetrieve\% (
  computeFunction() |>
    createJaspState()
)

result                         # the random number
\dontrun{isRecomputed(result)} # intentional error
isRecomputed(ctr[["state"]])   # true
isRecomputed()                 # true

result2 <- ctr[["state"]] \%setOrRetrieve\% (
  computeFunction() |>
    createJaspState()
)

identical(result, result2)   # the random number is identical, so retrieved from the state
isRecomputed(ctr[["state"]]) # false
isRecomputed()               # false
}
