% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformYeoJohnson.R
\name{YeoJohnson}
\alias{YeoJohnson}
\alias{YeoJohnsonAuto}
\alias{YeoJohnsonLambda}
\title{Yeo-Johnson transformation}
\usage{
YeoJohnson(x, lambda)

YeoJohnsonAuto(x, lower = -5, upper = 5)

YeoJohnsonLambda(x, lower = -5, upper = 5)
}
\arguments{
\item{x}{a numeric vector to be transformed.}

\item{lambda}{The lambda parameter of the transform.}

\item{lower}{Lower limit for possible lambda values.}

\item{upper}{Upper limit for possible lambda values.}
}
\value{
\code{YeoJohnson} and \code{YeoJohndonAuto} return the transformed variable.
\code{YeoJohnsonLambda} returns the optimal value of \code{lambda} for the given data.
}
\description{
Transform a variable using the Yeo-Johnson transformation that extends the Box-Cox transform to unbounded variables.
It is possible to select the value of \code{lambda} automatically by maximizing the normal profile log likelihood
}
\references{
Yeo, I. K., & Johnson, R. A. (2000). A new family of power transformations to improve normality or symmetry. \emph{Biometrika, 87}(4), 954-959.
}
