% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox.R
\name{BoxCox}
\alias{BoxCox}
\alias{invBoxCox}
\alias{BoxCoxAuto}
\alias{BoxCoxLambda}
\title{Box-Cox transformation}
\usage{
BoxCox(x, lambda, shift = 0)

invBoxCox(x, lambda, shift = 0)

BoxCoxAuto(
  x,
  method = c("guerrero", "loglik", "minitab"),
  lower,
  upper,
  shift = 0
)

BoxCoxLambda(x, method = c("guerrero", "loglik", "minitab"), lower, upper)
}
\arguments{
\item{x}{a numeric vector or time series of class \code{ts}.}

\item{lambda}{The lambda parameter of the transform.}

\item{shift}{The shift parameter of the transform. \code{shift=0} by default. All values of \code{x} must be larger than \code{shift}.}

\item{method}{Choose method to be used in calculating lambda. See details for explanation.}

\item{lower}{Lower limit for possible lambda values. By default set to -5 for \code{method=="minitab"}, otherwise -1.}

\item{upper}{Upper limit for possible lambda values. By default set to 5 for \code{method="minitab"}, otherwise 2.}
}
\value{
\code{BoxCox} and \code{BoxCoxAuto} return the transformed variable. \code{invBoxCox} returns the untransformed variable.
\code{BoxCoxLambda} returns the optimal value of \code{lambda} for the given data.
}
\description{
Transform a variable using the two-parameter Box-Cox transformation.
It is possible to select the value of \code{lambda} automatically
using one of three available methods.
}
\details{
If \code{method=="guerrero"}, Guerrero's (1993) method is used, where lambda
minimizes the coefficient of variation for subseries of \code{x}.

If \code{method=="loglik"}, the value of lambda is chosen to maximize the
profile log likelihood of a linear model fitted to \code{x}. For
non-seasonal data, a linear time trend is fitted while for seasonal data, a
linear time trend with seasonal dummy variables is used.

If \code{method=="minitab"}, the value of lambda is chosen to minimize
the sample standard deviation of the standardized, Box-Cox transformed variable.
This behavior mimics the auto Box-Cox transform available in the Minitab software.
}
\note{
This functionality is partially taken from the \code{forecast} package, which itself derived some code from the
package \code{MASS}. See also package \code{DescTools} for the Box-Cox transform.
}
\references{
Box, G. E. P. and Cox, D. R. (1964) An analysis of
transformations. \emph{JRSS B} \bold{26} 211--246.

Guerrero, V.M. (1993) Time-series analysis supported by power
transformations. \emph{Journal of Forecasting}, \bold{12}, 37--48.

https://support.minitab.com/en-us/minitab/help-and-how-to/quality-and-process-improvement/control-charts/how-to/box-cox-transformation/methods-and-formulas/methods-and-formulas/
}
\author{
Leanne Chhay and Rob J Hyndman and Šimon Kucharský
}
