% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformBoxCox.R
\name{BoxCox}
\alias{BoxCox}
\alias{invBoxCox}
\alias{BoxCoxAuto}
\alias{BoxCoxLambda}
\title{Box-Cox transformation}
\usage{
BoxCox(x, lambda, shift = 0, continuityAdjustment = TRUE)

invBoxCox(x, lambda, shift = 0, continuityAdjustment = TRUE)

BoxCoxAuto(
  x,
  predictor = NULL,
  groupSize = NULL,
  method = "loglik",
  lower = -5,
  upper = 5,
  shift = 0,
  continuityAdjustment = TRUE
)

BoxCoxLambda(
  x,
  predictor = NULL,
  groupSize = NULL,
  method = "loglik",
  lower = -5,
  upper = 5,
  shift = 0,
  continuityAdjustment = TRUE
)
}
\arguments{
\item{x}{a numeric vector.}

\item{lambda}{The lambda parameter of the transform.}

\item{shift}{The shift parameter of the transform. \code{shift=0} by default. Values of \code{x} smaller than \code{shift} will turn \code{NA}.}

\item{continuityAdjustment}{Boolean (default \code{TRUE}) whether to apply the continuity adjustment term (see Details).}

\item{predictor}{Vector indicating regression variable (factor or numeric).}

\item{groupSize}{Integer indicating the group size (exclusive argument with \code{predictor}). Must be larger than 1. For individual data, use \code{method="movingRange"}.}

\item{method}{Method for automatic determination of \code{lambda}. See details.}

\item{lower}{Lower limit for possible lambda values.}

\item{upper}{Upper limit for possible lambda values.}
}
\value{
\code{BoxCox} and \code{BoxCoxAuto} return the transformed variable. \code{invBoxCox} returns the untransformed variable.
\code{BoxCoxLambda} returns the optimal value of \code{lambda} for the given data.
}
\description{
Transform a variable using the two-parameter Box-Cox transformation.
It is possible to select the value of \code{lambda} automatically.
}
\details{
The two-parameter Box-Cox transformation is defined as
\deqn{y =
\begin{cases}
\frac{(x+\text{shift})^\lambda - 1}{\lambda} & \mathrm{if } \lambda \neq 0 \\
\log(x+\text{shift}) & \mathrm{if } \lambda = 0.
\end{cases}
}

If \code{continuityAdjustment=FALSE},
\deqn{y =
\begin{cases}
(x+\text{shift})^\lambda & \mathrm{if } \lambda \neq 0 \\
\log(x+\text{shift}) & \mathrm{if } \lambda = 0
\end{cases}
}

The continuity adjustment term ensures smooth convergence to the log transform as \eqn{\lambda \rightarrow 0}.

In \code{BoxCoxLambda} and \code{BoxCoxAuto}, the value of lambda is automatically selected using one of three possible methods:
\itemize{
\item When \code{method="loglik"}, lambda maximizes the normal log-likelihood of the transformed variable regressed on the predictor (Box & Cox, 1964, p. 215).
\item When \code{method="sd"}, lambda minimizes the residual sums of squares of the power-transformed variable regressed on the predictor (see \link{powerTransform}; Box & Cox, 1964, p. 216).
\item When \code{method="movingRange"}, lambda minimizes the estimate of variability based on the average moving range of the power-transformed variable (see Montgomerry, 2012).
Predictor is ignored in this case.
}

The predictor is either taken from the predictor argument, or inferred from the \code{groupSize} argument.
If both arguments are \code{NULL}, the data is assumed to be in one group (and regressed only on an intercept term).
}
\references{
Box, G. E. P. and Cox, D. R. (1964) An analysis of
transformations. \emph{JRSS B} \bold{26} 211--246.

Montgomery, D. C. (2012). Statistical quality control (Vol. 4). New York: John Wiley & Sons.
}
