% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformBoxCox.R
\name{BoxCox}
\alias{BoxCox}
\alias{invBoxCox}
\alias{BoxCoxAuto}
\alias{BoxCoxLambda}
\alias{BoxCoxMinitab}
\alias{BoxCoxLambdaMinitab}
\title{Box-Cox transformation}
\usage{
BoxCox(x, lambda, shift = 0)

invBoxCox(x, lambda, shift = 0)

BoxCoxAuto(x, lower = -5, upper = 5, shift = 0)

BoxCoxLambda(x, lower = -5, upper = 5, shift = 0)

BoxCoxMinitab(x, group = NULL, size = NULL, lower = -5, upper = 5)

BoxCoxLambdaMinitab(x, group = NULL, size = NULL, lower = -5, upper = 5)
}
\arguments{
\item{x}{a numeric vector.}

\item{lambda}{The lambda parameter of the transform.}

\item{shift}{The shift parameter of the transform. \code{shift=0} by default. All values of \code{x} must be larger than \code{shift}.}

\item{lower}{Lower limit for possible lambda values.}

\item{upper}{Upper limit for possible lambda values.}

\item{group}{Group vector indicating groups (for minitab emulation).}

\item{size}{If \code{group=NULL}, the number of observations per group.}
}
\value{
\code{BoxCox} and \code{BoxCoxAuto} return the transformed variable. \code{invBoxCox} returns the untransformed variable.
\code{BoxCoxLambda} returns the optimal value of \code{lambda} for the given data.
}
\description{
Transform a variable using the two-parameter Box-Cox transformation.
It is possible to select the value of \code{lambda} automatically using one of two available methods.
}
\details{
In \code{BoxCoxLambda} and \code{BoxCoxAuto}, the value of lambda is chosen to maximize the
normal profile log likelihood of the transformed variable.

The two-parameter Box-Cox transformation is defined as
\deqn{y =
\begin{cases}
\frac{(x+\text{shift})^\lambda - 1}{\lambda} & \mathrm{if } \lambda \neq 0 \\
\log(x+\text{shift}) & \mathrm{if } \lambda = 0.
\end{cases}
}
\subsection{Minitab emulation}{

\code{BoxCoxLambdaMinitab} and \code{BoxCoxMinitab} mimic the auto Box-Cox transform available in the Minitab software.
The value of lambda is chosen to minimize the pooled sample standard deviation
of the "standardized transformed variable" (see \link{powerTransform}), unless the data is based on individuals (ungrouped),
that is, when \code{size=1} or all values in \code{group} are unique: then lambda minimizes the
unbiased estimate of the average moving range.
The data is then transformed using a simple exponential equation (neglecting the normalizing factor of the standard Box-Cox transform)
\deqn{y =
\begin{cases}
x^\lambda & \mathrm{if } \lambda \neq 0 \\
\log(x) & \mathrm{if } \lambda = 0
\end{cases}
}
}
}
\references{
Box, G. E. P. and Cox, D. R. (1964) An analysis of
transformations. \emph{JRSS B} \bold{26} 211--246.

\href{https://support.minitab.com/en-us/minitab/help-and-how-to/quality-and-process-improvement/control-charts/how-to/box-cox-transformation/methods-and-formulas/methods-and-formulas/}{Box-Cox transformation in Minitab}

\href{https://support.minitab.com/en-us/minitab/help-and-how-to/quality-and-process-improvement/control-charts/how-to/variables-charts-for-subgroups/i-mr-r-s-chart/methods-and-formulas/estimating-sigma-for-the-i-chart-and-the-mr-chart/#average-moving-range-method}{Unbiased average moving range in Minitab}
}
