% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-types.R
\name{column-types}
\alias{column-types}
\alias{jaspScale}
\alias{isJaspScale}
\alias{asJaspScale}
\title{JASP Column Types}
\usage{
jaspScale(x = double())

isJaspScale(x)

asJaspScale(x, ...)
}
\arguments{
\item{x}{the main object for which the operation is done}
}
\description{
Columns types in JASP.
}
