% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleInstall.R
\name{installModuleNew}
\alias{installModuleNew}
\title{Install a Jasp module and update its dependencies.}
\usage{
installModuleNew(
  modulePath,
  jaspRoot,
  moduleLibrary,
  lockfilePath = NULL,
  updatePackages = Sys.getenv("JASP_UPDATE_PKGS", unset = "false"),
  recordPackages = c("localJasp", "all"),
  recurseJaspDependencies = TRUE,
  useLocalLockfile = FALSE,
  respectModuleLockfile = TRUE,
  prompt = FALSE,
  verbose = 1L
)
}
\arguments{
\item{modulePath}{local path to jasp module, e.g., ~/github/jasp-desktop/Modules/jaspDescriptives (note that this does not need to be inside jasp-desktop)}

\item{jaspRoot}{local path to clone of jasp-desktop, e.g., ~/github/jasp-desktop}

\item{moduleLibrary}{where to install the module}

\item{lockfilePath}{optional, path to a lockfile.}

\item{updatePackages}{optional, whether to update the non JASP R package dependencies. Defaults to false, but see Details.}

\item{recordPackages}{optional, whether to record all package dependencies as they are installed ("all"), or whether to change records of Jasp modules to the locally checked out version ("localJasp").}

\item{recurseJaspDependencies}{optional, whether to use the local versions of indirect Jasp module dependencies. Defaults to TRUE.}

\item{useLocalLockfile}{optional, whether to use the lockfile present in modulePath. Defaults to false.}

\item{prompt}{optional, show renv promts. Ignored when not running in interactive mode (e.g., inside Jasp).}
}
\value{
returns \code{NULL}.
}
\description{
Install a Jasp module and update its dependencies.
}
\details{
\code{installModuleNew} does four things:
\enumerate{
\item Install a Jasp module from scratch using locally checked out jasp modules. When \code{recurseJaspDependencies} is TRUE (the default), then it is ensured that any indirect dependencies on other jasp modules als use the locally checked out versions.
\item Update the Jasp module dependencies when locally checked out sources change. For example, when jaspBase is updated but jaspDescriptives not, this function ensures that jaspDescriptives will use the new version of jaspBase.
\item Optionally, update non Jasp R package dependencies.
\item TODO: Optionally, install a Jasp module using the lockfile specified in the source location. This ignores any options set for the previous three options.
}

Possible values for \code{updatePackages} are "true", "false", "daily", "triweekly", "biweekly", "weekly", "fortnightly", "monthly",
"monday", "tuesday", ... "sunday", or a combination of weekdays, "monday;wednesday;friday".
Anything value that is not understood is interpreted as FALSE (with a warning).
The date of the previous update is determined by looking at the last date where \code{lockfilePath} was modified.
}
